--- webdev/react/Name_Surname/src/components/Footer.jsx
@@ -1,13 +1,0 @@
-import React from 'react'
-
-const Footer = () => {
-  return (
-    <div className='footer'>
-        ©️Newsify@2022
-        <br />
-        Contact Us: info@newsify.com
-    </div>
-  )
-}
-
-export default Footer
