--- webdev/react/Name_Surname/src/components/Politics.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Politics = () => {
-  return (
-    <div>Politics</div>
-  )
-}
-
-export default Politics
