--- webdev/react/Name_Surname/src/components/Sports.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Sports = () => {
-  return (
-    <div>Sports</div>
-  )
-}
-
-export default Sports
