+++ webdev/react/Tarun/src/Social.jsx
@@ -1,0 +1,17 @@
+import React from 'react'
+
+const Social = () => {
+  return (
+    <div>
+        <h4>Connect with us at</h4>
+    <div className='social'>
+    <a href="#"><i class="fa-brands fa-facebook"></i></a>
+    <a href="#"><i class="fa-brands fa-twitter"></i></a>
+    <a href="#"><i class="fa-brands fa-youtube"></i></a>
+    <a href="#"><i class="fa-brands fa-instagram"></i></a>
+    </div>    
+    </div>
+  )
+}
+
+export default Social
