+++ webdev/react/Tarun/src/components/Footer.jsx
@@ -1,0 +1,13 @@
+import React from 'react'
+
+const Footer = () => {
+  return (
+    <div className='footer'>
+        ©️Newsify@2022
+        <br />
+        Contact Us: info@newsify.com
+    </div>
+  )
+}
+
+export default Footer
