+++ webdev/react/Tarun/src/components/Politics.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Politics = () => {
+  return (
+    <div>Politics</div>
+  )
+}
+
+export default Politics
