+++ webdev/react/Tarun/src/components/Sports.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Sports = () => {
+  return (
+    <div>Sports</div>
+  )
+}
+
+export default Sports
