+++ webdev/react/Tarun/src/components/Top.jsx
@@ -1,0 +1,13 @@
+import React from 'react'
+
+function Top() {
+  return (
+    <header>
+        <img src={require('../images/logo.png')} alt="" className='logo' />
+        <h1 className='page_heading'>Newsify</h1>
+        
+    </header>
+  )
+}
+
+export default Top
