+++ webdev/react/Tarun/src/index.js
@@ -1,0 +1,13 @@
+import React from "react";
+import ReactDOM  from "react-dom";
+import App from "./App";
+import {BrowserRouter} from 'react-router-dom'
+
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(
+    <React.StrictMode>
+        <BrowserRouter>
+            <App />
+        </BrowserRouter>
+    </React.StrictMode>
+);
